requirement.txt
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
idna==3.11
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.2.1
narwhals==2.9.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pick==2.4.0
plotly==6.3.1
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
